pyproject.toml
[tool.poetry]
name = "api-template"
version = "0.1.0"
description = "The API template for the TCE-PA Projects"
authors = ["SETIN"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.112.0"
uvicorn = "0.30.6"
python-dotenv = "1.0.1"
supabase = "2.7.3"
sentry-sdk = { extras = ["fastapi"], version = "2.13.0" }
requests-sse = "0.3.2"
pydantic-settings = "2.4.0"
redis = "5.0.8"
pika = "1.3.2"
pika-stubs = "0.1.3"
pyyaml = "^6.0.2"
celery = {extras = ["redis"], version = "^5.4.0"}
crontab = "^1.0.1"
psutil = "^6.0.0"
pytest-cov = "^5.0.0"
flower = "^2.0.1"
passlib = "^1.7.4"
jose = "^1.0.0"
alembic = "^1.13.2"
pytest = "^8.3.2"
aio-pika = "^9.4.3"
circuitbreaker = "^2.0.0"
python-jose = "^3.3.0"
psycopg2-binary = "^2.9.9"
pydantic = {extras = ["email"], version = "^2.8.2"}
cachetools = "^5.5.0"
frozendict = "^2.4.4"
openai = "^1.44.0"
qdrant-client = "^1.11.1"
langfuse = "^2.47.0"
marvin = "^2.3.7"
semver = "^3.0.2"

[tool.poetry.dev-dependencies]
pytest = "8.3.2"
pytest-mock = "3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100

[tool.flake8]
max-line-length = 100
extend-ignore = "E203, W503, E501"
exclude = ".git, __pycache__, .venv, .eggs, *.egg, build, dist"

[tool.autopep8]
max_line_length = 100
in-place = true
recursive = true
aggressive = 3

[tool.poetry.scripts]
migrate = "scripts/run_migration.sh"
